refactor: moving the cors_settings into the web package
