+++ azurerm/internal/services/web/cors_settings.go
@@ -1,4 +1,4 @@
-package azure
+package web
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/web/mgmt/2019-08-01/web"
